+++ app/src/main/java/com/example/myapplication/ui/dashboard/DashboardViewModel.java
@@ -1,19 +1,19 @@
-package com.example.myapplication.ui.dashboard;
-
-import androidx.lifecycle.LiveData;
-import androidx.lifecycle.MutableLiveData;
-import androidx.lifecycle.ViewModel;
-
-public class DashboardViewModel extends ViewModel {
-
-    private final MutableLiveData<String> mText;
-
-    public DashboardViewModel() {
-        mText = new MutableLiveData<>();
-        mText.setValue("This is dashboard fragment");
-    }
-
-    public LiveData<String> getText() {
-        return mText;
-    }
-}
+//package com.example.myapplication.ui.dashboard;
+//
+//import androidx.lifecycle.LiveData;
+//import androidx.lifecycle.MutableLiveData;
+//import androidx.lifecycle.ViewModel;
+//
+//public class DashboardViewModel extends ViewModel {
+//
+//    private final MutableLiveData<String> mText;
+//
+//    public DashboardViewModel() {
+//        mText = new MutableLiveData<>();
+//        mText.setValue("This is dashboard fragment");
+//    }
+//
+//    public LiveData<String> getText() {
+//        return mText;
+//    }
+//}
