+++ app/src/main/java/com/example/myapplication/ui/feedback/FeedbackViewModel.java
@@ -1,0 +1,7 @@
+package com.example.myapplication.ui.feedback;
+
+import androidx.lifecycle.ViewModel;
+
+public class FeedbackViewModel extends ViewModel {
+    // TODO: Implement the ViewModel
+}
