+++ app/src/main/java/com/example/myapplication/ui/search/SearchViewModel.java
@@ -1,0 +1,4 @@
+package com.example.myapplication.ui.search;
+
+public class SearchViewModel {
+}
