+++ app/src/main/res/layout/fragment_home.xml
@@ -7,7 +7,7 @@
     tools:context=".ui.home.HomeFragment">
 
     <include
-        layout="@layout/top_nav"
+        layout="@layout/topbar"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:layout_alignParentBottom="true" />
