+++ app/src/main/res/layout/fragment_search.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent">
 
     <include
-        layout="@layout/top_nav"
+        layout="@layout/topbar"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:layout_alignParentBottom="true" />
@@ -19,10 +19,11 @@
         app:layout_constraintGuide_begin="214dp" />
 
     <SearchView
-        android:layout_width="337dp"
+        android:layout_width="0dp"
         android:layout_height="51dp"
+        android:layout_marginTop="152dp"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
-        tools:layout_editor_absoluteY="151dp" />
+        app:layout_constraintTop_toTopOf="parent" />
 
 </androidx.constraintlayout.widget.ConstraintLayout>
