+++ app/src/main/res/menu/bottom_nav_menu.xml
@@ -1,5 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
-<menu xmlns:android="http://schemas.android.com/apk/res/android">
+<menu xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:android="http://schemas.android.com/apk/res/android">
 
     <item
         android:id="@+id/navigation_home"
@@ -7,13 +8,24 @@
         android:title="@string/title_home" />
 
     <item
-        android:id="@+id/navigation_dashboard"
-        android:icon="@drawable/ic_dashboard_black_24dp"
-        android:title="@string/title_dashboard" />
-
-    <item
         android:id="@+id/navigation_notifications"
         android:icon="@drawable/ic_notifications_black_24dp"
         android:title="@string/title_notifications" />
+
+    <item
+        android:id="@+id/navigation_search"
+        android:icon="@drawable/ic_search_black_24dp"
+        android:title="@string/title_search"
+        app:actionViewClass="android.widget.SearchView" />
+
+    <item
+        android:id="@+id/navigation_feedback"
+        android:icon="@drawable/ic_dashboard_black_24dp"
+        android:title="@string/title_feedback" />
+
+    <item
+        android:id="@+id/navigation_profile"
+        android:icon="?android:attr/listChoiceIndicatorSingle"
+        android:title="@string/title_profile" />
 
 </menu>
