+++ app/src/main/res/values/strings.xml
@@ -3,4 +3,7 @@
     <string name="title_home">Home</string>
     <string name="title_dashboard">Dashboard</string>
     <string name="title_notifications">Notifications</string>
+    <string name="title_search">Search</string>
+    <string name="title_feedback">Feedback</string>
+    <string name="title_profile">Profile</string>
 </resources>
